+++ NEWS.md
@@ -11,9 +11,11 @@
 ## Added
 - New `check_py_for_scanpy()` function, shifting code around in `anndataToGiotto()`
 - Add `initialize()` method for `exprObj` to automatically format matrix info
+- Add `instructions()` generic for `giotto` to access and edit `giottoInstructions`
 - Add `centroids()` method for `giottoPolygon` to get centroid info
 - Add `overlaps()` generic for accessing `overlaps` slot of `giottoPolygon`
 - Add `[` and `[<-` access generics to main slots of `giottoPolygon` and `giottoPoints`
+- Add cores detection to run on package attach
 - New `dd.R` as a file with dummy documentation to inherit commonly used documentation from
 - Add option to return as `giottoPoints` from `getFeatureInfo` (default is still `SpatVector`)
 - New `assign_objnames_2_list()` and `assign_listnames_2_obj()` internals for passing list names to object `@name` slots and vice versa